Support non ascii characters in Ifmap server
When the ifmap server recieved a message, the envelope is
recieved as a 'String', where as when the ifmap client
calculates the length of the envelope, the type of the
envelope is 'unicode'. There is a mismatch in the length
if the envelop contains non ascii characters.

A new patch is added for the ifmap client which returns the
the 'envelope' as a 'str' type.

Please see the bug description for details

Change-Id: I59b20971c27b82f4cb3bdbef2346505fc0feceed
Closes-bug: #1413946
(cherry picked from commit 5f69cac6d4d049b2459bc0f18830fe3e5b5f4f2d)