[vrouter_netns] Create netns if it's not exist
If the script opencontrail-vrouter-netns is call with the '--uptae'
flag, the Linux network namespace is not initialized before it sets SNAT
or LBaaS stuff on it. So it fails. To resolve that, it create it if it's
not exist.

Conflicts:
	src/vnsw/opencontrail-vrouter-netns/opencontrail_vrouter_netns/vrouter_netns.py

Change-Id: I74f87f3e7f1ebca1ac664d646d139a4f6eddbf98
Closes-Bug: #1466879
(cherry picked from commit cc61f98cdfd5f289ef3adeab4d3f756d1912fa0a)
(cherry picked from commit e9bc752958e595c98af930ab1a9ef90e6d553fc4)