Allow to use 80/90 for service instance local pref
This is a workaround regarding the leak of snat default route to the
public vrf.

Before this patch, the agent only supported local pref 100 and 200. It
now also supports local pref 80 and 90. By doing this, even if a leak
occurs, the default route annonced by the border router (pref 100)
will have a higer priority than routes that leaked from the snat.

We continue to support local_pref 100 and 200 since there are already
created SI with these local prefs.

Note the snat that leaks doens't work anymore.

Change-Id: Ic9911582d5408a57c2aa845b1d68d66e5852173f